requirements.txt
sentence-transformers
python-docx
faiss-cpu
transformers
fastapi
uvicorn
pydantic